+++ supercollider/synth.py
@@ -63,7 +63,7 @@ class Synth(object):
         self.server._send_msg("/s_get", self.id, parameter)
 
         if blocking:
-            rv = self.server._await_response("/n_set", [self.id, parameter])
+            rv = self.server._await_response("/n_set", [self.id, parameter], self.server.param_handler)
             return rv
         else:
             self.server._add_handler("/n_set", [self.id, parameter], callback)
